backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.0
Pillow==10.2.0  # For image handling
django-cors-headers==4.3.1  # For handling cross-origin requests when working with Next.js
python-decouple==3.8  # For environment variables
whitenoise==6.6.0  # For serving static files
django-filter==24.1  # For filtering in API 